+++ BlueprintUICommonControls/Tests/Sources/AttributedLabelTests.swift
@@ -357,6 +357,87 @@ class AttributedLabelTests: XCTestCase {
         compareSnapshot(of: element)
     }
 
+    func test_linkAccessibility() {
+        let labelview = AttributedLabel.LabelView()
+
+        do {
+            // Test that link insertion happy path works
+            let string = NSString("Foo Bar Baz")
+            let url = URL(string: "https://block.xyz")!
+            for (word, result) in [
+                ("Foo", "Foo[Link] Bar Baz"),
+                ("Bar", "Foo Bar[Link] Baz"),
+                ("Baz", "Foo Bar Baz[Link]"),
+            ] {
+                let range = string.range(of: word)
+                let link = AttributedLabel.Link(url: url, range: range)
+                let accessibilityLabel = labelview.accessibilityLabel(
+                    with: [link],
+                    in: string as String,
+                    linkAccessibilityLabel: "Link"
+                )
+                XCTAssertEqual(accessibilityLabel, result)
+            }
+        }
+
+        do {
+            // Test every position
+            let numbers = [1, 2, 3, 4, 5, 6, 7, 8, 9, 10]
+            let string = numbers.map { String($0) }.joined() as NSString
+            let url = URL(string: "https://block.xyz")!
+            for number in numbers {
+                let range = NSMakeRange(max(0, number - 2), 1)
+                let link = AttributedLabel.Link(url: url, range: range)
+                let accessibilityLabel = labelview.accessibilityLabel(
+                    with: [link],
+                    in: string as String,
+                    linkAccessibilityLabel: "."
+                ) as NSString
+                XCTAssertNotEqual(accessibilityLabel, string)
+            }
+        }
+
+        do {
+            // Test stupid ranges don't crash
+            let string = "Foo Bar Baz"
+            let url = URL(string: "https://block.xyz")!
+            let badRanges = [
+                NSMakeRange(0, 0),
+                NSMakeRange(-1, 0),
+                NSMakeRange(0, -1),
+                NSMakeRange(0, 100),
+                NSMakeRange(100, 0),
+                NSMakeRange(100, -100),
+            ]
+            for range in badRanges {
+                let link = AttributedLabel.Link(url: url, range: range)
+                let accessibilityLabel = labelview.accessibilityLabel(
+                    with: [link],
+                    in: string as String,
+                    linkAccessibilityLabel: "."
+                )
+                XCTAssertEqual(accessibilityLabel, string)
+            }
+        }
+
+        do {
+            // Test with emoji
+            let string = "🇺🇸🇨🇦🇯🇵🇫🇷"
+            let url = URL(string: "https://block.xyz")!
+            let range = NSRange(string.range(of: "🇨🇦")!, in: string)
+            let link = AttributedLabel.Link(url: url, range: range)
+
+            let accessibilityLabel = labelview.accessibilityLabel(
+                with: [link],
+                in: string as String,
+                linkAccessibilityLabel: "."
+            )
+            XCTAssertEqual(accessibilityLabel, "🇺🇸🇨🇦[.]🇯🇵🇫🇷")
+        }
+    }
+
+
+
     func test_textContainerRects() {
         let lineBreakModes: [NSLineBreakMode?] = [
             nil,